backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
boto3>=1.34.0
requests>=2.32.0
python-dotenv>=1.0.1
pydantic>=2.9.0
playwright>=1.48.0
